tui: Add Orders page, trading shortcuts, and order entry dialog (#145)
<img width="597" height="319" alt="image"
src="https://github.com/user-attachments/assets/43129df1-783e-4fb7-9cbf-442c3dd28635"
/>

<img width="1491" height="215" alt="image"
src="https://github.com/user-attachments/assets/0f3ad756-544d-4d18-b046-864b403bc70d"
/>

🤖 Auto-generated by Endless.

---------

+++ src/tui/render/dirty_flags.rs
@@ -80,7 +80,7 @@ impl DirtyFlags {
     /// Mark components for a popup change
     #[inline]
     #[must_use]
-    pub fn mark_popup_change(mut self, popup: u8) -> Self {
+    pub fn mark_popup_change(mut self, popup: u16) -> Self {
         if popup & crate::tui::app::POPUP_HELP != 0 {
             self.insert(Self::POPUP_HELP);
         }
